+++ src/components/MovieInfo/MovieInfo.scss
