+++ src/components/MovieInfo/__tests__/movieInfo.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ShallowRenderer from 'react-test-renderer/shallow';
-import MovieDetailsInfo from '../MovieDetailsInfo';
+import MovieDetailsInfo from '../MovieInfo';
 
 const info = {
   title: 'title',
