+++ src/components/MovieRating/MovieRating.scss
